+++ x/tieredrewards/keeper/msg_validate.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/crypto-org-chain/chain-main/v8/x/tieredrewards/types"
 
+	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/math"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -207,6 +208,39 @@ func (k Keeper) validateWithdrawFromTier(ctx context.Context, pos types.Position
 	}
 	if unbonding {
 		return types.ErrPositionUnbonding
+	}
+
+	return nil
+}
+
+func (k Keeper) validateExitTierWithDelegation(ctx context.Context, pos types.Position, owner string, amount math.Int) error {
+	if !pos.IsOwner(owner) {
+		return types.ErrNotPositionOwner
+	}
+
+	if !pos.IsDelegated() {
+		return types.ErrPositionNotDelegated
+	}
+
+	if !pos.HasTriggeredExit() {
+		return types.ErrExitNotTriggered
+	}
+
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	if !pos.CompletedExitLockDuration(sdkCtx.BlockTime()) {
+		return types.ErrExitLockDurationNotReached
+	}
+
+	if amount.GT(pos.Amount) {
+		return errorsmod.Wrapf(types.ErrInvalidAmount, "amount %s exceeds position amount %s", amount, pos.Amount)
+	}
+
+	redelegating, err := k.stillRedelegating(ctx, pos.Id)
+	if err != nil {
+		return err
+	}
+	if redelegating {
+		return errorsmod.Wrapf(types.ErrActiveRedelegation, "position %d has an active redelegation", pos.Id)
 	}
 
 	return nil